+++ gridpad/src/main/kotlin/com/touchlane/gridpad/GridPadSpanAnchor.kt
@@ -1,0 +1,14 @@
+package com.touchlane.gridpad
+
+/**
+ * Anchor for spanned cells
+ */
+internal data class GridPadSpanAnchor(val horizontal: Horizontal, val vertical: Vertical) {
+    enum class Horizontal {
+        START, END
+    }
+
+    enum class Vertical {
+        TOP, BOTTOM
+    }
+}