+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Change Log
 ==========
 
+## Version 1.0.0
+
+_2022-12-28_
+
+* **New**: added `GridPadPlacementPolicy` class that used to control implicit placement policy
+  for items
+* **New**: added `GridPadDiagnosticLogger` diagnostic class to receive information about skipped
+  items
+* **API Change**: to `GridPad` added `placementPolicy` property
+* **API Change**: `GridPadScope.item()` split to two methods - explicit and implicit
+* **Behavior Change**: implicit placement of elements depends on `placementPolicy` property
+* **Behavior Change**: span expands depends on `placementPolicy` property
+
+The new API shouldn't affect your code by default. Only one case can be affected by library 
+update - cases when only one of the `row` or `column` parameters was defined in GridPad.item. 
+This way of using is no more acceptable for not enough clear behavior reasons.
+
 ## Version 0.0.4
 
 _2022-12-20_